chore: reorganize project directory structure
