+++ core/database/consumer-rules.pro
@@ -1,3 +1,3 @@
 -keep class com.waffiq.bazz_movies.core.database.data.model.** { *; }
 -keep class com.waffiq.bazz_movies.core.database.utils.DbResult { *; }
--keep class com.waffiq.bazz_movies.core.database.domain.usecase.local_database.** { *; }
+-keep class com.waffiq.bazz_movies.core.database.domain.usecase.localdatabase.** { *; }
