+++ core/uihelper/src/main/kotlin/com/waffiq/bazz_movies/core/uihelper/utils/ActionBarBehavior.kt
@@ -10,6 +10,15 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.updateLayoutParams
 import com.google.android.material.appbar.AppBarLayout
 
+/**
+ * A utility object that provides functions for modifying the appearance and behavior of the status bar
+ * and handling app bar layout in an Android application.
+ *
+ * This object includes:
+ * - [transparentStatusBar], methods to make the status bar transparent
+ * - [handleOverHeightAppBar], to adjust the app bar layout to handle system insets, ensuring proper
+ *   display on different device configurations.
+ */
 object ActionBarBehavior {
 
   @Suppress("DEPRECATION")
